requirements.txt
pandas==1.3.5
fastapi==0.79.0
python-dotenv==0.20.0
pathlib
pydantic==1.9.0
uvicorn==0.18.2
Jinja2==3.1.2